pyproject.toml
[project]
name = "paradox-script-mcp"
version = "0.1.0"
description = "MCP server for efficient Paradox game script discovery"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "paradox-script-parser",
    "pyyaml>=6.0.3",
]

[project.scripts]
paradox-script-mcp = "paradox_script_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/paradox_script_mcp"]

[tool.uv.sources]
paradox-script-parser = { git = "https://github.com/106-/paradox-script-parser" }

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]
